Remove low-value tests: constant checks, examples, benchmarks, duplicates
Removed 8 test files (772 lines):
- cluster/cluster_timeout_test.go — trivial constant value assertions
- cluster/config_numshards_test.go — trivial config field assignment check
- gate/gateserver/gate_timeout_unit_test.go — default value checks + Go stdlib behavior
- cluster/cluster_ready_verification_test.go — duplicate of cluster_ready_test.go
- cluster/cluster_ready_nodeconnections_test.go — duplicate of cluster_ready_test.go
- util/testutil/example_dynamic_ports_test.go — demo/example, no assertions
- util/workerpool/benchmark_test.go — perf comparison, not functional test
- util/workerpool/efficiency_test.go — goroutine count comparison
